requirements.txt
click==8.0.1
numpy==1.22.3
pandas==1.4.2
kafka-python==2.0.2
requests==2.27.1
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0